==> code/01_data preparation and mcdm implementation.Rmd <==
---
title: "Data Preparation and MCDM Implementation"
# output: html_notebook
author:
  - name: Qiuping Ma
date: "`r Sys.Date()`"
# creative_commons: CC BY-NC
output:
  html_notebook:
    toc: true
    toc_depth: 3
    number_sections: true
    theme: paper
    highlight: tango
    chunk_output_type: inline
---
# Introduction
In this R document, we execute R code for supplier quality evaluation. To be more specific, we work on data preparation and MCDM methods implementation. Six MCDM methods are implemented, including TOPSIS, VIKOR, PROMETHEE II--Usual type, PROMETHEE II--Gaussian type, MOORA--The ration system, and MOORA--The reference approach.

# Preparation
%Given an evaluation matrix with 100 suppliers against 23 criteria, we removed the criteria C21, since 85% percent of its value is 0.

## Set working directory
For operation simplicity, we set the working directory as below.
```{r}
setwd("C:/Users/au630301/Documents/00.Qiuping/2nd Paper/code")
getwd()
```
## Load the package
```{r eval = FALSE}
if(!require("tidyverse")) install.packages("tidyverse")
library(tidyverse)
if(!require("MCDA")) install.packages("MCDA")
library("MCDA")
if(!require("MCDM")) install.packages("MCDM")
library("MCDM")
```

## Data set import and preprocessing
Now we can import the data set and have a quick look at the data set.
```{r eval = FALSE}
dat <- read_csv("2018_11_supplier evaluation original data_final.csv")
```
```{r}
dat[22,]
```

```{r}
head(dat,3)
```
```{r}
str(dat)
```
After the quick look, we start to preprocess the data set for different data types, so that it can be used for different MCDM models. Normalization is used to eliminate the units of criterion functions, so that all the criteria are dimensionless. According to the requirements of the six MCDM models, we scale the criteria values such that the column has euclidian length 1. 
Non-beneficial criteria include C09, C18-C23. Others are beneficial criteria.

The linear normalization formulas are $x_{ij}^* = \frac{x-x_{min}}{x_{max}-x_{min}}$ for beneficial criteria and $x_{ij}^* = \frac{x_{max}-x}{x_{max}-x_{min}}$ for non-beneficial criteria. For every criterion, the minimum value of that criterion gets transformed into a 0, the maximum value gets transformed into a 1, and every other value gets transformed into a decimal between 0 and 1.
The vector normalization formula is $x_{ij}^* = \frac{x_{ij}}{ \sqrt[]{\sum_{i=1}^{m}  x_{ij} ^{2} } }$. 

According to MCDM methods requirement, different normalization procedures are choosed. VIKOR, PROMETHEE II--Usual type, PROMETHEE II--Gaussian type use linear normalization. The others use vector normalization. 
Since we have to use 2 normalization procedures, 2 normalized datasets (dat_norm_linear  and dat_norm_vector ) will be generated. 

```{r}
class(dat)
```

```{r}
# for beneficial criteria
norm_minmax <- function(x){
  (x- min(x)) /(max(x)-min(x))
}
```
```{r}
# for non-beneficial criteria
norm_minmax_2 <- function(x){
  (x- max(x)) /(min(x)-max(x))
}
```
```{r}
# for beneficial criteria
dat_norm_linear_1 <- apply(dat[,-c(1,10,19,20,21,22,23,24)], 2, norm_minmax)
#dat_norm_linear_1
```
```{r}
# for non-beneficial criteria
dat_norm_linear_2 <- apply(dat[, c(10,19,20,21,22,23,24)], 2, norm_minmax_2)
#dat_norm_linear_2
```

```{r}
dat_norm_linear_initial <- cbind(dat_norm_linear_1, dat_norm_linear_2)
#dat_norm_linear_initial
dat_norm_linear <- dat_norm_linear_initial[, c(1:8,17,9,10,11,12,13,14,15,16,18:23 )]
dat_norm_linear
#write_csv(as.data.frame(All_result_norm_new), "processed3/All_result_norm_new.csv" )
```

```{r}
normalizationTypes <- rep("scaleToUnitLength",ncol(dat)-1)
dat_norm_vector <- normalizePerformanceTable(performanceTable = dat[,-1], normalizationTypes)
dat_norm_vector
```


```{r}
write_csv(as.data.frame(dat_norm_linear), "processed4/dat_norm_linear.csv" )
write_csv(dat_norm_vector, "processed4/dat_norm_vector.csv" )
```
# MCDM method implementation
Since we have four dimensions, we create four data sets accordingly, so that we are able to evaluate performance for each dimension.

## Preparation
```{r}
dat_linear_sr <- as.data.frame(dat_norm_linear) %>%
  select(C01:C05)
head(dat_linear_sr,3)

dat_vector_sr <- dat_norm_vector %>%
  select(C01:C05)
head(dat_vector_sr,3)
```
```{r}
dat_linear_pi <- as.data.frame(dat_norm_linear) %>%
  select(C06:C10)
head(dat_linear_pi,3)

dat_vector_pi <- dat_norm_vector %>%
  select(C06:C10)
head(dat_vector_pi,3)
```
```{r}
dat_linear_sw <- as.data.frame(dat_norm_linear) %>%
  select(C11:C17)
head(dat_linear_sw,3)

dat_vector_sw <- dat_norm_vector %>%
  select(C11:C17)
head(dat_vector_sw,3)
```

```{r}
dat_linear_sqc <- as.data.frame(dat_norm_linear) %>%
  select(C18:C23)
head(dat_linear_sqc,3)

dat_vector_sqc <- dat_norm_vector %>%
  select(C18:C23)
head(dat_vector_sqc,3)
```
## Criteria weight setting
Generally, decision-makers provide the weight of criteria based on their preferences. Since criteria weight has no influence on the illusion of the method application, we treat each criterion equally important. 
```{r}
weights_sr <- rep(1/ncol(dat_vector_sr), ncol(dat_vector_sr))
weights_pi <- rep(1/ncol(dat_vector_pi), ncol(dat_vector_pi))
weights_sw <- rep(1/ncol(dat_vector_sw), ncol(dat_vector_sw))
weights_sqc <- rep(1/ncol(dat_vector_sqc), ncol(dat_vector_sqc))
weights_sr
weights_pi
weights_sw
weights_sqc
```
## Benefit and cost (bc) setting
```{r}
bc_sr  <- c("max", "max", "max", "max", "max")
bc_pi  <- c("max", "max", "max", "min", "max")
bc_sw  <- c("max", "max", "max", "max", "max", "max", "max")
bc_sqc <- c("min", "min", "min", "min", "min", "min")

```

## TOPSIS method
It uses vector normalization.
### Supply risk
```{r}
criteriaMinMax_topsis_sr <- bc_sr
names(weights_sr) <- colnames(dat_vector_sr)
names(criteriaMinMax_topsis_sr) <- colnames(dat_vector_sr)
overall_sr <- TOPSIS(dat_vector_sr, weights_sr, criteriaMinMax_topsis_sr)
str(overall_sr)
class(overall_sr)
```
### Profit impact
```{r}
criteriaMinMax_topsis_pi <- bc_pi
names(weights_pi) <- colnames(dat_vector_pi)
names(criteriaMinMax_topsis_pi) <- colnames(dat_vector_pi)
overall_pi <- TOPSIS(dat_vector_pi, weights_pi, criteriaMinMax_topsis_pi)
str(overall_pi)
class(overall_pi)
```
### Supplier willingness
```{r}
criteriaMinMax_topsis_sw <- bc_sw
names(weights_sw) <- colnames(dat_vector_sw)
names(criteriaMinMax_topsis_sw) <- colnames(dat_vector_sw)
overall_sw <- TOPSIS(dat_vector_sw, weights_sw, criteriaMinMax_topsis_sw)
str(overall_sw)
class(overall_sw)
```
### Supplier quality capability
```{r}
criteriaMinMax_topsis_sqc <- bc_sqc
names(weights_sqc) <- colnames(dat_vector_sqc)
names(criteriaMinMax_topsis_sqc) <- colnames(dat_vector_sqc)
overall_sqc <- TOPSIS(dat_vector_sqc, weights_sqc, criteriaMinMax_topsis_sqc)
str(overall_sqc)
```
### Result summary
```{r}
TOPSIS_result <- matrix(, nrow = nrow(dat), ncol = 4)
TOPSIS_result[,1] <- matrix(overall_sr)
TOPSIS_result[,2] <- matrix(overall_pi)
TOPSIS_result[,3] <- matrix(overall_sw)
TOPSIS_result[,4] <- matrix(overall_sqc)
```
```{r}
head(TOPSIS_result,5)
```
## VIKOR method
It uses linear normalization. The smaller the better. 
### Supply risk
```{r}
v_sr <- 0.65
VIKOR_sr <- VIKOR(data.matrix(dat_linear_sr, rownames.force = NA), weights_sr, bc_sr, v_sr)
#VIKOR_sr
```
```{r}
v_pi <- 0.65
VIKOR_pi <- VIKOR(data.matrix(dat_linear_pi, rownames.force = NA), weights_pi, bc_pi, v_pi)
#VIKOR_pi
```
```{r}
v_sw <- 0.65
VIKOR_sw <- VIKOR(data.matrix(dat_linear_sw, rownames.force = NA), weights_sw, bc_sw, v_sw)
#VIKOR_sw
```
```{r}
v_sqc <- 0.65
VIKOR_sqc <- VIKOR(data.matrix(dat_linear_sqc, rownames.force = NA), weights_sqc, bc_sqc, v_sqc)
#VIKOR_sqc
```
### Result summary
```{r}
VIKOR_result <- matrix(, nrow = nrow(dat), ncol = 4)
VIKOR_result[,1] <- matrix(VIKOR_sr[,4])
VIKOR_result[,2] <- matrix(VIKOR_pi[,4])
VIKOR_result[,3] <- matrix(VIKOR_sw[,4])
VIKOR_result[,4] <- matrix(VIKOR_sqc[,4])
```
```{r}
head(VIKOR_result,5)
```
## PROMETHEE Method with Usual preference function
It uses linear normalization. 
Note that if we want to use RPrometheeConstructor() function, the number of criteria should be less than or equal to 5. 
```{r}
if(!require("RMCriteria")) install.packages("RMCriteria")
library("RMCriteria")
```
### Supply risk
```{r}
parms_sr <- matrix(rep(NA, ncol(dat_linear_sr)), byrow = TRUE, ncol=1, nrow=ncol(dat_linear_sr))
vecWeights_sr <- weights_sr
vecMaximiz_sr <- c(T, T, T, T, T)
preferenceF_sr_up <- rep(1, ncol(dat_linear_sr))
PROMETHEE_up_sr <- PrometheeII(datMat = as.matrix(dat_linear_sr), vecWeights = vecWeights_sr, prefFunction = preferenceF_sr_up, parms = parms_sr, normalize = FALSE)
```
### Profit impact
```{r}
parms_pi <- matrix(rep(NA, ncol(dat_linear_pi)), byrow = TRUE, ncol=1, nrow=ncol(dat_linear_pi))
vecWeights_pi <- weights_pi
vecMaximiz_pi <- c(T, T, T, F, T)
preferenceF_pi_up <- rep(1, ncol(dat_linear_pi))
PROMETHEE_up_pi <- PrometheeII(datMat = as.matrix(dat_linear_pi), vecWeights = vecWeights_pi, prefFunction = preferenceF_pi_up, parms = parms_pi, normalize = FALSE)
```

### Supplier willingness
Note that since we have more than 5 criteria, we have to use PrometheeII() function.
```{r}
parms_sw <- matrix(rep(NA, ncol(dat_linear_sw)), byrow = TRUE, ncol=1, nrow=ncol(dat_linear_sw))
vecWeights_sw <- weights_sw
vecMaximiz_sw <- c(T, T, T, T, T, T, T)
preferenceF_sw_up <- rep(1, ncol(dat_linear_sw))
PROMETHEE_up_sw <- PrometheeII(datMat = as.matrix(dat_linear_sw), vecWeights = vecWeights_sw, prefFunction = preferenceF_sw_up, parms = parms_sw, normalize = FALSE)
```
### Supplier quality capacity
```{r}
parms_sqc <- matrix(rep(NA, ncol(dat_linear_sqc)), byrow = TRUE, ncol=1, nrow=ncol(dat_linear_sqc))
vecWeights_sqc <- weights_sqc
vecMaximiz_sqc <- c(F, F, F, F, F, F)
preferenceF_sqc_up <- rep(1, ncol(dat_linear_sqc))
PROMETHEE_up_sqc <- PrometheeII(datMat = as.matrix(dat_linear_sqc), vecWeights = vecWeights_sqc, prefFunction = preferenceF_sqc_up, parms = parms_sqc, normalize = FALSE)
```
### Result summary
```{r}
PROMETHEE_result_up <- matrix(, nrow = nrow(dat), ncol = 4)
PROMETHEE_result_up[,1] <- matrix(PROMETHEE_up_sr)
PROMETHEE_result_up[,2] <- matrix(PROMETHEE_up_pi)
PROMETHEE_result_up[,3] <- matrix(PROMETHEE_up_sw)
PROMETHEE_result_up[,4] <- matrix(PROMETHEE_up_sqc)
```
```{r}
head(PROMETHEE_result_up,5)
```
## PROMETHEE Method with Gaussian preference function. 
Note that if we want to use RPrometheeConstructor() function, the number of criteria should be less than or equal to 5. 
```{r}
if(!require("RMCriteria")) install.packages("RMCriteria")
library("RMCriteria")
```
### Supply risk
```{r}
preferenceF_gp_sr <- rep(0, ncol(dat_linear_sr))
PROMETHEE_gp_sr <- PrometheeII(datMat = as.matrix(dat_linear_sr), vecWeights = vecWeights_sr, prefFunction = preferenceF_gp_sr, parms = parms_sr, normalize = FALSE)
```
### Profit impact
```{r}
preferenceF_gp_pi <- rep(0, ncol(dat_linear_pi))
PROMETHEE_gp_pi <- PrometheeII(datMat = as.matrix(dat_linear_pi), vecWeights = vecWeights_pi, prefFunction = preferenceF_gp_pi, parms = parms_pi, normalize = FALSE)
```
### Supplier willingness
Note that since we have more than 5 criteria, we have to use PrometheeII() function.
```{r}
preferenceF_gp_sw <- rep(0, ncol(dat_linear_sw))
PROMETHEE_gp_sw <- PrometheeII(datMat = as.matrix(dat_linear_sw), vecWeights = vecWeights_sw, prefFunction = preferenceF_gp_sw, parms = parms_sw, normalize = FALSE)
```
### Supplier quality capacity
```{r}
preferenceF_gp_sqc <- rep(0, ncol(dat_linear_sqc))
PROMETHEE_gp_sqc <- PrometheeII(datMat = as.matrix(dat_linear_sqc), vecWeights = vecWeights_sqc, prefFunction = preferenceF_gp_sqc, parms = parms_sqc, normalize = FALSE)
```
### Result summary
```{r}
PROMETHEE_result_gp <- matrix(, nrow = nrow(dat), ncol = 4)
PROMETHEE_result_gp[,1] <- matrix(PROMETHEE_gp_sr)
PROMETHEE_result_gp[,2] <- matrix(PROMETHEE_gp_pi)
PROMETHEE_result_gp[,3] <- matrix(PROMETHEE_gp_sw)
PROMETHEE_result_gp[,4] <- matrix(PROMETHEE_gp_sqc)
```
```{r}
head(PROMETHEE_result_gp,5)
```

## MOORA method
### Supply risk
```{r}
MMOORA_sr <- MMOORA(as.matrix(dat_vector_sr),weights_sr,bc_sr)
MMOORA_sr_rs <- matrix(MMOORA_sr$RatioSystem)
MMOORA_sr_rp <- matrix(MMOORA_sr$ReferencePoint)
```
### Profit impact
```{r}
MMOORA_pi <- MMOORA(as.matrix(dat_vector_pi),weights_pi,bc_pi)
MMOORA_pi_rs <- matrix(MMOORA_pi$RatioSystem)
MMOORA_pi_rp <- matrix(MMOORA_pi$ReferencePoint)
```
### Supplier willingness
```{r}
MMOORA_sw <- MMOORA(as.matrix(dat_vector_sw),weights_sw,bc_sw)
MMOORA_sw_rs <- matrix(MMOORA_sw$RatioSystem)
MMOORA_sw_rp <- matrix(MMOORA_sw$ReferencePoint)
```
### Supplier quality capability
```{r}
MMOORA_sqc <- MMOORA(as.matrix(dat_vector_sqc),weights_sqc,bc_sqc)
MMOORA_sqc_rs <- matrix(MMOORA_sqc$RatioSystem)
MMOORA_sqc_rp <- matrix(MMOORA_sqc$ReferencePoint)
```
### Result summary
```{r}
MMOORA_rs_result <- matrix(, nrow = nrow(dat), ncol = 4)
MMOORA_rs_result[,1] <- MMOORA_sr_rs
MMOORA_rs_result[,2] <- MMOORA_pi_rs
MMOORA_rs_result[,3] <- MMOORA_sw_rs
MMOORA_rs_result[,4] <- MMOORA_sqc_rs
```
```{r}
MMOORA_rp_result <- matrix(, nrow = nrow(dat), ncol = 4)
MMOORA_rp_result[,1] <- MMOORA_sr_rp
MMOORA_rp_result[,2] <- MMOORA_pi_rp
MMOORA_rp_result[,3] <- MMOORA_sw_rp
MMOORA_rp_result[,4] <- MMOORA_sqc_rp
```

```{r}
MMOORA_sqc
```
## MCDM implementation result summary, processing, and saving
### Result summary
```{r}
All_result <- matrix(, nrow = nrow(dat), ncol = 24)
All_result <- cbind(TOPSIS_result,VIKOR_result,PROMETHEE_result_up, PROMETHEE_result_gp, MMOORA_rs_result, MMOORA_rp_result)
MCDM_result <- All_result[,c(1,5,9,13,17,21,2,6,10,14,18,22,3,7,11,15,19,23,4,8,12,16,20,24)]
colnames(MCDM_result) <- c("M1_sr", "M2_sr", "M3_sr", "M4_sr", "M5_sr","M6_sr","M1_pi", "M2_pi", "M3_pi", "M4_pi", "M5_pi", "M6_pi","M1_sw", "M2_sw", "M3_sw", "M4_sw", "M5_sw","M6_sw", "M1_sqc", "M2_sqc", "M3_sqc", "M4_sqc", "M5_sqc","M6_sqc")
write_csv(as.data.frame(MCDM_result), "processed4/MCDM_result.csv" )
```
### Processing
For aggregation purposes, linear normalization is performed on the MCDM_result data set. Again, we shold separate beneficial score and non-beneficial score. TOPSIS, PROMETHEE-Usual type, PROMETHEE-Gaussian type, and MOORA-Ratio system have beneficial scores and the others get non-beneficial scores.
```{r}
# for beneficial score
MCDM_result_1 <- apply(MCDM_result[,-c(2, 6, 8, 12, 14, 18, 24)], 2, norm_minmax)
#MCDM_result_1
```
```{r}
# for non-beneficial score -- column 2, 6, 8, 12, 14, 18, 20, 24
MCDM_result_2 <- apply(MCDM_result[,c(2, 6, 8, 12, 14, 18, 24)], 2, norm_minmax_2)
#MCDM_result_2
```
```{r}
MCDM_result_3 <- cbind(MCDM_result_1, MCDM_result_2)
#MCDM_result_3
MCDM_result_norm <- MCDM_result_3[, c(1,18,2,3,4,19, 5,20,6,7,8,21, 9,22,10,11,12,23, 13,14,15,16,17,24)]
write_csv(as.data.frame(MCDM_result_norm), "processed4/MCDM_result_norm.csv" )
```
# MCDM method comparison
Given the results of six MCDM methods, now we use ggplot2 package to have a look at the results. Totally, we have 100 suppliers. For simplicity, we plot for the first 10 suppliers as example. We have four dimensions, thus four plots should be generated. We install the ggplot2 package. Add a new column with supplier ID.

## Preparation
```{r}
if(!require("ggplot2")) install.packages("ggplot2")
library("ggplot2")
if(!require("repr")) install.packages("repr")
library("repr")
```
```{r}
Supplier_ID <- c("A_{01}", "A_{02}", "A_{03}", "A_{04}", "A_{05}","A_{06}", "A_{07}", "A_{08}", "A_{09}", "A_{10}",
                 "A_{11}", "A_{12}", "A_{13}", "A_{14}", "A_{15}","A_{16}", "A_{17}", "A_{18}", "A_{19}", "A_{20}",
                 "A_{21}", "A_{22}", "A_{23}", "A_{24}", "A_{25}","A_{26}", "A_{27}", "A_{28}", "A_{29}", "A_{30}",
                 "A_{31}", "A_{32}", "A_{33}", "A_{34}", "A_{35}","A_{36}", "A_{37}", "A_{38}", "A_{39}", "A_{40}",
                 "A_{41}", "A_{42}", "A_{43}", "A_{44}", "A_{45}","A_{46}", "A_{47}", "A_{48}", "A_{49}", "A_{50}",
                 "A_{51}", "A_{52}", "A_{53}", "A_{54}", "A_{55}","A_{56}", "A_{57}", "A_{58}", "A_{59}", "A_{60}",
                 "A_{61}", "A_{62}", "A_{63}", "A_{64}", "A_{65}","A_{66}", "A_{67}", "A_{68}", "A_{69}", "A_{70}",
                 "A_{71}", "A_{72}", "A_{73}", "A_{74}", "A_{75}","A_{76}", "A_{77}", "A_{78}", "A_{79}", "A_{80}",
                 "A_{81}", "A_{82}", "A_{83}", "A_{84}", "A_{85}","A_{86}", "A_{87}", "A_{88}", "A_{89}", "A_{90}",
                 "A_{91}", "A_{92}", "A_{93}", "A_{94}", "A_{95}","A_{96}", "A_{97}", "A_{98}", "A_{99}", "A_{100}"
)  

MCDM_result_norm_new<- cbind(Supplier_ID, as.data.frame(MCDM_result_norm))
write_csv(as.data.frame(MCDM_result_norm_new), "processed4/MCDM_result_norm_new.csv" )
```
## Plot

### Supply risk
```{r}
if(!require("reshape2")) install.packages("reshape2")
library("reshape2")
```
```{r}
df_sr <- MCDM_result_norm_new[1:10, 1:7] %>%
  rename(M1 = M1_sr,
         M2 = M2_sr,
         M3 = M3_sr,
         M4 = M4_sr,
         M5 = M5_sr,
         M6 = M6_sr
  )

df_sr_melt = melt(df_sr, id = "Supplier_ID")
```
```{r}
df_sr_melt <-df_sr_melt %>%
  rename(Method = variable,
         Score = value)
df_sr_melt
```
```{r}
plot_sr <- ggplot(df_sr_melt, aes(x = Method, y = Score, group = Supplier_ID)) + 
  geom_line() + 
  facet_wrap(Supplier_ID ~ ., nrow = 2, ncol = 5) +
  geom_point() +
  labs(
  title = "Supplier risk",
  x = "models",
  y = "Score"
  ) +
theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
  ylim(0, 1)
ggsave(file="processed4/plot/plot_sr.pdf",width=8, height=2.5, dpi=300)
```
### Profit impact
```{r}
df_pi <- MCDM_result_norm_new[1:10, c(1,8:13)] %>%
  rename(M1 = M1_pi,
         M2 = M2_pi,
         M3 = M3_pi,
         M4 = M4_pi,
         M5 = M5_pi,
         M6 = M6_pi
  )

df_pi_melt = melt(df_pi, id = "Supplier_ID")
```
```{r}
df_pi_melt <-df_pi_melt %>%
  rename(Method = variable,
         Score = value)

df_pi_melt
```
```{r}
plot_pi <- ggplot(df_pi_melt, aes(x = Method, y = Score, group = Supplier_ID)) + 
  geom_line() + 
  facet_wrap(Supplier_ID ~ ., nrow = 2, ncol = 5) +
  geom_point() +
  labs(
  title = "Profit impact",
  x = "models",
  y = "Score"
  ) +
theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
  ylim(0, 1)
ggsave(file="processed4/plot/plot_pi.pdf",width=8, height=2.5, dpi=300)
```
### Supplier willingness
```{r}
df_sw <- MCDM_result_norm_new[1:10, c(1,14:19)] %>%
  rename(M1 = M1_sw,
         M2 = M2_sw,
         M3 = M3_sw,
         M4 = M4_sw,
         M5 = M5_sw,
         M6 = M6_sw
         )

df_sw_melt = melt(df_sw, id = "Supplier_ID")
```
```{r}
df_sw_melt <-df_sw_melt %>%
  rename(Method = variable,
         Score = value)

df_sw_melt
```
```{r}
plot_sw <- ggplot(df_sw_melt, aes(x = Method, y = Score, group = Supplier_ID)) + 
  geom_line() + 
  facet_wrap(Supplier_ID ~ ., nrow = 2, ncol = 5) +
  geom_point() +
  labs(
  title = "Supplier willingness",
  x = "models",
  y = "Score"
  ) +
theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
  ylim(0, 1)
ggsave(file="processed4/plot/plot_sw.pdf",width=8, height=2.5, dpi=300)
```
### Supplier quality capability
```{r}
df_sqc <- MCDM_result_norm_new[1:10, c(1,20:25)] %>%
  rename(M1 = M1_sqc,
         M2 = M2_sqc,
         M3 = M3_sqc,
         M4 = M4_sqc,
         M5 = M5_sqc,
         M6 = M6_sqc,
         )

df_sqc_melt = melt(df_sqc, id = "Supplier_ID")
```
```{r}
df_sqc_melt <-df_sqc_melt %>%
  rename(Method = variable,
         Score = value)

df_sqc_melt
```
```{r}
plot_sc <- ggplot(df_sqc_melt, aes(x = Method, y = Score, group = Supplier_ID)) + 
  geom_line() + 
  facet_wrap(Supplier_ID ~ ., nrow = 2, ncol = 5) +
  geom_point() +
  labs(
  title = "Supplier quality capability",
  x = "models",
  y = "Score"
  ) +
theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) + 
  ylim(0, 1)
ggsave(file="processed4/plot/plot_sc.pdf",width=8, height=2.5, dpi=300)
```
### Plot summary
```{r}
if(!require("ggpubr")) install.packages("ggpubr")
library("ggpubr")
```
```{r}
ggarrange(plot_sr,plot_pi, plot_sw, plot_sc,
          # labels = c("PPM_Q", "PPM_M","SPM_Q", "SPM_M"),
          ncol = 1, nrow = 4)
ggsave(file="processed4/plot/plot_summary.pdf",width=8, height=10, dpi=300)
```
## Similarity calculation

As we can see the plot of supply risk, MMOORA_Reference method seems to have a big deviation when compared with other methods. Thus, we assume that MMOORA_Reference method may have the lowest similarity with other methods, To justify our assumption, we calculate the euclidean distance between each method.

### Supply risk
```{r}
dist_sr <- dist(t(MCDM_result_norm_new[, 2:7]), upper = TRUE)
```
```{r}
#Calculate the mean distance.
Ave_dist_sr<- rowMeans(as.matrix(dist_sr))
Ave_dist_sr
```

### Profit impact
```{r}
dist_pi <- dist(t(MCDM_result_norm_new[, 8:13]), upper = TRUE)
```
```{r}
#Calculate the mean distance.
Ave_dist_pi<- rowMeans(as.matrix(dist_pi))
Ave_dist_pi
```
### Supplier willingness
```{r}
dist_sw <- dist(t(MCDM_result_norm_new[, 14:19]), upper = TRUE)
dist_sw
```
```{r}
#Calculate the mean distance.
Ave_dist_sw<- rowMeans(as.matrix(dist_sw))
Ave_dist_sw
```

### Supplier quality capability 
```{r}
dist_sqc <- dist(t(MCDM_result_norm_new[, 20:25]), upper = TRUE)
dist_sqc
```
```{r}
#Calculate the mean distance.
Ave_dist_sqc<- rowMeans(as.matrix(dist_sqc))
Ave_dist_sqc
```

### Similarity summary
Now we make a summary of the similarity between different MCDM methods. For the four dimensions, the results may vary. In the dimension of supply risk and supplier willingness, Model 5 (MOORA with the reference approach) shows the greatest difference. While in the dimension of profit impact and supplier quality capability, Model 2 (PROMETHEE with Usual type) shows the biggest gap. Two conclusions can be drawn here: (i) different models may yield different performance given different datasets; (ii) an aggregation model is essential to obtain a comprise solution based on multiple MCDM models.
```{r}
Ave_dist <- rbind(Ave_dist_sr, Ave_dist_pi, Ave_dist_sw, Ave_dist_sqc) %>%
  as.data.frame() %>%
    rename(
      TOPSIS = M1_sr,
      VIKOR = M2_sr,
      PROMETHEE_Usual = M3_sr,
      PROMETHEE_Gaussian = M4_sr,
      MMOORA_Ratio = M5_sr,
      MMOORA_Reference= M6_sr
  )
Ave_dist 
```
```{r}
write_csv(as.data.frame(Ave_dist), "processed4/Ave_dist.csv" )
```
